@@ -16,8 +16,8 @@ import io.vertx.ext.web.RoutingContext;
 /**
  * A handler for requests that do not match the operations defined in the fester OpenAPI specification.
  * <p>
- * This handler can fail the context with any response code other than a 404. Because it's set up to catch 404s,
- * failing the routing context with a 404 will create a loop. To return a legitimate 404 from this handler, use
+ * This handler can fail the context with any response code other than a 404. Because it's set up to catch 404s, failing
+ * the routing context with a 404 will create a loop. To return a legitimate 404 from this handler, use
  * HttpServerReponse and set 404 as the response code, remembering to call <code>.end()</code> on the response.
  */
 public class MatchingOpNotFoundHandler implements Handler<RoutingContext> {
